+++ tests/assets_manager_test.cc
@@ -31,6 +31,7 @@ TEST(AssetsManager, file_image) {
 
   r = assets_manager_ref(rm, ASSET_TYPE_IMAGE, "earth");
   ASSERT_EQ(r != NULL, true);
+  ASSERT_EQ(r->refcount, 1);
   ASSERT_EQ(assets_manager_unref(rm, r), RET_OK);
 
   assets_manager_destroy(rm);
@@ -42,7 +43,14 @@ TEST(AssetsManager, file_script) {
 
   r = assets_manager_ref(rm, ASSET_TYPE_SCRIPT, "dummy");
   ASSERT_EQ(r != NULL, true);
+  ASSERT_EQ(r->refcount, 2);
+
+  r = assets_manager_find_in_cache(rm, ASSET_TYPE_SCRIPT, "dummy");
+  ASSERT_EQ(r != NULL, true);
+
   ASSERT_EQ(assets_manager_unref(rm, r), RET_OK);
+  r = assets_manager_find_in_cache(rm, ASSET_TYPE_SCRIPT, "dummy");
+  ASSERT_EQ(r->refcount, 1);
 
   assets_manager_destroy(rm);
 }