+++ tests/style_mutable_test.cc
@@ -22,6 +22,8 @@ static ret_t on_style_item(void* ctx, const char* widget_state, const char* id,
 }
 
 TEST(StyleMutable, basic) {
+  value_t v;
+  value_t v1;
   int32_t i = 0;
   int32_t k = 0;
   int32_t n = 100;
@@ -53,6 +55,18 @@ TEST(StyleMutable, basic) {
       snprintf(font_name, sizeof(font_name), "font%d", i);
       ASSERT_EQ(style_mutable_set_str(style, state, STYLE_ID_FONT_NAME, font_name), RET_OK);
       ASSERT_EQ(string(style_get_str(style, STYLE_ID_FONT_NAME, "")), string(font_name));
+
+      snprintf(font_name, sizeof(font_name), "fonta%d", i);
+      value_set_str(&v, font_name);
+      ASSERT_EQ(style_mutable_set_value(style, state, STYLE_ID_FONT_NAME, &v), RET_OK);
+
+      ASSERT_EQ(string(style_get_str(style, STYLE_ID_FONT_NAME, "")), string(font_name));
+      ASSERT_EQ(style_mutable_get_value(style, state, STYLE_ID_FONT_NAME, &v1), RET_OK);
+      ASSERT_STREQ(value_str(&v1), font_name);
+
+      snprintf(font_name, sizeof(font_name), "font%d", i);
+      ASSERT_EQ(style_mutable_set_str(style, state, STYLE_ID_FONT_NAME, font_name), RET_OK);
+      ASSERT_EQ(string(style_get_str(style, STYLE_ID_FONT_NAME, "")), string(font_name));
     }
   }
 
@@ -64,5 +78,16 @@ TEST(StyleMutable, basic) {
             "over,fg_color,65536;over,font_name,\"font99\";");
 
   style_destroy(style);
+  widget_destroy(w);
+}
+
+TEST(StyleMutable, cast) {
+  widget_t* w = window_create(NULL, 10, 20, 30, 40);
+  widget_t* b = button_create(w, 0, 0, 100, 100);
+  style_t* style_mutable = style_mutable_create(b);
+
+  ASSERT_EQ(style_mutable, style_mutable_cast(style_mutable));
+
+  style_destroy(style_mutable);
   widget_destroy(w);
 }